+++ Docs/CHANGELOG.md
@@ -11,6 +11,9 @@
 * OpenSCENARIO support:
     - Added support for position with Lane information  (roadId and laneId)
     - Added support for TimeOfDay tag
+* Scenario updates:
+    - Scenarios that are part of RouteScenario have had their triggering condition modified. This will only activate when a certain parameter is set, and if not, the old trigger condition will still be applied.
+
 ### :bug: Bug Fixes
 * Fixed spawning bugs for scenario DynamicObjectCrossing when it is part of a route
 * Fixed spawning bugs for scenarios VehicleTurningRight, VehicleTurningLeft when they are part of a route